split off serverApi into its own file for clearer separation
